split up tests into files
