+++ test/services/api/handlers/docs/index.js
@@ -8,7 +8,7 @@ var Lab = require('lab'),
   server;
 
 before(function(done) {
-  require('../../../mocks/server')(function(obj) {
+  require('../../../../mocks/server')(function(obj) {
     server = obj;
     done();
   });
