+++ test/services/api/handlers/pages/delete.js
@@ -1,0 +1,202 @@
+var configs = require('../../../../fixtures/configs/page-handlers'),
+  sinon = require('sinon'),
+  Lab = require('lab'),
+  lab = exports.lab = Lab.script(),
+  experiment = lab.experiment,
+  before = lab.before,
+  after = lab.after,
+  test = lab.test,
+  expect = require('code').expect,
+  server,
+  thumbnailServiceUrl;
+
+function mockErr() {
+  var e = new Error('relation does not exist');
+  e.name = 'error';
+  e.severity = 'ERROR';
+  e.code = '42P01';
+  return e;
+}
+
+before(function(done) {
+  thumbnailServiceUrl = process.env.THUMBNAIL_SERVICE_URL;
+  delete process.env.THUMBNAIL_SERVICE_URL;
+  require('../../../../mocks/server')(function(obj) {
+    server = obj;
+    done();
+  });
+});
+
+after(function(done) {
+  process.env.THUMBNAIL_SERVICE_URL = thumbnailServiceUrl;
+  server.stop(done);
+});
+
+experiment('DELETE /users/{user}/projects/{project}/pages/{page}', function() {
+  test('success - owner', function(done) {
+    var opts = configs.del.success.owner;
+
+    server.inject(opts, function(resp) {
+      expect(resp.statusCode).to.equal(200);
+      expect(resp.result.status).to.equal('deleted');
+      server.inject(
+        '/users/1/projects/1/pages/2',
+        function(resp) {
+          expect(resp.statusCode).to.equal(404);
+          done();
+        }
+      );
+    });
+  });
+
+  test('success - moderator', function(done) {
+    var opts = configs.del.success.moderator;
+
+    server.inject(opts, function(resp) {
+      expect(resp.statusCode).to.equal(200);
+      expect(resp.result.status).to.equal('deleted');
+      server.inject(
+        '/users/1/projects/1/pages/3',
+        function(resp) {
+          expect(resp.statusCode).to.equal(404);
+          done();
+        }
+      );
+    });
+  });
+
+  test('user not found', function(done) {
+    var opts = configs.del.fail.params.user.notFound;
+
+    server.inject(opts, function(resp) {
+      expect(resp.statusCode).to.equal(404);
+      expect(resp.result.error).to.equal('Not Found');
+      expect(resp.result.message).to.equal('User not found');
+      done();
+    });
+  });
+
+  test('user not number', function(done) {
+    var opts = configs.del.fail.params.user.notNumber;
+
+    server.inject(opts, function(resp) {
+      expect(resp.statusCode).to.equal(400);
+      expect(resp.result.error).to.equal('Bad Request');
+      expect(resp.result.message).to.be.a.string();
+      done();
+    });
+  });
+
+  test('user not integer', function(done) {
+    var opts = configs.del.fail.params.user.notInteger;
+
+    server.inject(opts, function(resp) {
+      expect(resp.statusCode).to.equal(400);
+      expect(resp.result.error).to.equal('Bad Request');
+      expect(resp.result.message).to.be.a.string();
+      done();
+    });
+  });
+
+  test('user does not own project', function(done) {
+    var opts = configs.del.fail.params.user.doesNotOwnProject;
+
+    server.inject(opts, function(resp) {
+      expect(resp.statusCode).to.equal(404);
+      expect(resp.result.error).to.equal('Not Found');
+      expect(resp.result.message).to.be.a.string();
+      done();
+    });
+  });
+
+  test('project not found', function(done) {
+    var opts = configs.del.fail.params.project.notFound;
+
+    server.inject(opts, function(resp) {
+      expect(resp.statusCode).to.equal(404);
+      expect(resp.result.error).to.equal('Not Found');
+      expect(resp.result.message).to.equal('Project not found');
+      done();
+    });
+  });
+
+  test('project not number', function(done) {
+    var opts = configs.del.fail.params.project.notNumber;
+
+    server.inject(opts, function(resp) {
+      expect(resp.statusCode).to.equal(400);
+      expect(resp.result.error).to.equal('Bad Request');
+      expect(resp.result.message).to.be.a.string();
+      done();
+    });
+  });
+
+  test('project not integer', function(done) {
+    var opts = configs.del.fail.params.project.notInteger;
+
+    server.inject(opts, function(resp) {
+      expect(resp.statusCode).to.equal(400);
+      expect(resp.result.error).to.equal('Bad Request');
+      expect(resp.result.message).to.be.a.string();
+      done();
+    });
+  });
+
+  test('page is not part of project', function(done) {
+    var opts = configs.del.fail.params.project.pageNotInProject;
+
+    server.inject(opts, function(resp) {
+      expect(resp.statusCode).to.equal(404);
+      expect(resp.result.error).to.equal('Not Found');
+      expect(resp.result.message).to.be.a.string();
+      done();
+    });
+  });
+
+  test('page not found', function(done) {
+    var opts = configs.del.fail.params.page.notFound;
+
+    server.inject(opts, function(resp) {
+      expect(resp.statusCode).to.equal(404);
+      expect(resp.result.error).to.equal('Not Found');
+      expect(resp.result.message).to.equal('Page not found');
+      done();
+    });
+  });
+
+  test('page not number', function(done) {
+    var opts = configs.del.fail.params.page.notNumber;
+
+    server.inject(opts, function(resp) {
+      expect(resp.statusCode).to.equal(400);
+      expect(resp.result.error).to.equal('Bad Request');
+      expect(resp.result.message).to.be.a.string();
+      done();
+    });
+  });
+
+  test('page not integer', function(done) {
+    var opts = configs.del.fail.params.page.notInteger;
+
+    server.inject(opts, function(resp) {
+      expect(resp.statusCode).to.equal(400);
+      expect(resp.result.error).to.equal('Bad Request');
+      expect(resp.result.message).to.be.a.string();
+      done();
+    });
+  });
+
+  test('Handles errors from postgre', function(done) {
+    var opts = configs.del.fail.error;
+    var stub = sinon.stub(server.methods.pages, 'remove')
+      .callsArgWith(1, mockErr());
+
+    server.inject(opts, function(resp) {
+      expect(resp.statusCode).to.equal(500);
+      expect(resp.result.error).to.equal('Internal Server Error');
+      expect(resp.result.message).to.be.a.string();
+      stub.restore();
+      done();
+    });
+  });
+});
