+++ test/services/api/handlers/projects/prerequisites.js
@@ -1,0 +1,59 @@
+var configs = require('../../../../fixtures/configs/project-handlers'),
+  sinon = require('sinon'),
+  Lab = require('lab'),
+  lab = exports.lab = Lab.script(),
+  experiment = lab.experiment,
+  before = lab.before,
+  after = lab.after,
+  test = lab.test,
+  expect = require('code').expect,
+  server;
+
+function mockErr() {
+  var e = new Error('relation does not exist');
+  e.name = 'error';
+  e.severity = 'ERROR';
+  e.code = '42P01';
+  return e;
+}
+
+before(function(done) {
+  require('../../../../mocks/server')(function(obj) {
+    server = obj;
+    done();
+  });
+});
+
+after(function(done) {
+  server.stop(done);
+});
+
+experiment('Project prequisite errors', function() {
+  test('getUser pg error', function(done) {
+    var opts = configs.prerequisites.fail;
+    var stub = sinon.stub(server.methods.users, 'find')
+      .callsArgWith(1, mockErr());
+
+    server.inject(opts, function(resp) {
+      expect(resp.statusCode).to.equal(500);
+      expect(resp.result.error).to.equal('Internal Server Error');
+      expect(resp.result.message).to.equal('An internal server error occurred');
+      stub.restore();
+      done();
+    });
+  });
+
+  test('getProject pg error', function(done) {
+    var opts = configs.prerequisites.fail;
+    var stub = sinon.stub(server.methods.projects, 'findOne')
+      .callsArgWith(1, mockErr());
+
+    server.inject(opts, function(resp) {
+      expect(resp.statusCode).to.equal(500);
+      expect(resp.result.error).to.equal('Internal Server Error');
+      expect(resp.result.message).to.equal('An internal server error occurred');
+      stub.restore();
+      done();
+    });
+  });
+});
